Adjust base layout styles
