+++ gatsby-theme-landing-page/src/components/base.module.css
@@ -54,7 +54,10 @@
 } */
 
 .container {
+  width: 100%;
   max-width: var(--max-width);
   padding-left: var(--space-4);
   padding-right: var(--space-4);
+  margin-left: auto;
+  margin-right: auto;
 }
